+++ pht/helpers/ProxyActions.sol
@@ -25,7 +25,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("rely(address,address)", from, to);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -38,7 +38,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("deny(address,address)", from, to);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -51,7 +51,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("file(address,bytes32,uint256)", who, what, data);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -64,7 +64,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("file(address,bytes32,bytes32,uint256)", who, ilk, what, data);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -77,7 +77,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("dripAndFile(address,bytes32,uint256)", who, what, data);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -89,14 +89,8 @@ contract ProxyActions is DSAuth {
         assembly {
             tag := extcodehash(usr)
         }
-        bytes memory fax = abi.encodeWithSignature(
-            "dripAndFile(address,bytes32,bytes32,uint256)",
-            who,
-            ilk,
-            what,
-            data
-        );
-        uint eta = now;
+        bytes memory fax = abi.encodeWithSignature("dripAndFile(address,bytes32,bytes32,uint256)", who, ilk, what, data);
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -109,7 +103,7 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("cage(address)", end);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
@@ -122,38 +116,34 @@ contract ProxyActions is DSAuth {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("setAuthority(address,address)", pause, newAuthority);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
     }
 
-    function setDelay(uint newDelay) external auth {
+    function setDelay(uint256 newDelay) external auth {
         address usr = address(govActions);
         bytes32 tag;
         assembly {
             tag := extcodehash(usr)
         }
         bytes memory fax = abi.encodeWithSignature("setDelay(address,uint256)", pause, newDelay);
-        uint eta = now;
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
     }
 
-    function setAuthorityAndDelay(address newAuthority, uint newDelay) external auth {
+    function setAuthorityAndDelay(address newAuthority, uint256 newDelay) external auth {
         address usr = address(govActions);
         bytes32 tag;
         assembly {
             tag := extcodehash(usr)
         }
-        bytes memory fax = abi.encodeWithSignature(
-            "setAuthorityAndDelay(address,address,uint256)",
-            pause,
-            newAuthority,
-            newDelay
-        );
-        uint eta = now;
+        bytes memory fax =
+            abi.encodeWithSignature("setAuthorityAndDelay(address,address,uint256)", pause, newAuthority, newDelay);
+        uint256 eta = now;
 
         pause.plot(usr, tag, fax, eta);
         pause.exec(usr, tag, fax, eta);
